refacto: déplace et remplace des fixtures spécifiques au module api les api recherche entreprises, sirene et ratios financiers ne sont pas appelées directement par notre code mais toujours via l'api interne infos_entreprises qui sert d'interface
