+++ impact/conftest.py
@@ -143,21 +143,22 @@ def entreprise_factory(db, date_cloture_dernier_exercice):
 # (entreprise, reglementations, users ...), hence pulled-up.
 
 
-INFOS_ENTREPRISE = {
-    "siren": "000000001",
-    "denomination": "Entreprise SAS",
-    "effectif": CaracteristiquesAnnuelles.EFFECTIF_MOINS_DE_10,
-    "categorie_juridique_sirene": 5710,
-    "code_pays_etranger_sirene": None,
-    "code_NAF": "01.11Z",
-}
-
-
 @pytest.fixture
 def mock_api_infos_entreprise(mocker):
+    infos_entreprises_avec_donnees_financieres = {
+        "siren": "000000001",
+        "denomination": "Entreprise SAS",
+        "effectif": CaracteristiquesAnnuelles.EFFECTIF_MOINS_DE_10,
+        "categorie_juridique_sirene": 5710,
+        "code_pays_etranger_sirene": None,
+        "code_NAF": "01.11Z",
+        "date_cloture_exercice": date(2023, 12, 31),
+        "tranche_chiffre_affaires": CaracteristiquesAnnuelles.CA_MOINS_DE_900K,
+        "tranche_chiffre_affaires_consolide": CaracteristiquesAnnuelles.CA_MOINS_DE_60M,
+    }
     return mocker.patch(
         "api.infos_entreprise.infos_entreprise",
-        return_value=INFOS_ENTREPRISE,
+        return_value=infos_entreprises_avec_donnees_financieres,
     )
 
 
@@ -171,34 +172,6 @@ def mock_api_recherche_par_nom_ou_siren(mocker):
     return mocker.patch(
         "api.infos_entreprise.recherche_par_nom_ou_siren",
         return_value=entreprises,
-    )
-
-
-@pytest.fixture
-def mock_api_recherche_entreprises(mocker):
-    return mocker.patch(
-        "api.recherche_entreprises.recherche_par_siren",
-        return_value=INFOS_ENTREPRISE,
-    )
-
-
-@pytest.fixture
-def mock_api_sirene(mocker):
-    return mocker.patch(
-        "api.sirene.recherche_unite_legale",
-        return_value=INFOS_ENTREPRISE,
-    )
-
-
-@pytest.fixture
-def mock_api_ratios_financiers(mocker):
-    return mocker.patch(
-        "api.ratios_financiers.dernier_exercice_comptable",
-        return_value={
-            "date_cloture_exercice": date(2023, 12, 31),
-            "tranche_chiffre_affaires": CaracteristiquesAnnuelles.CA_MOINS_DE_900K,
-            "tranche_chiffre_affaires_consolide": CaracteristiquesAnnuelles.CA_MOINS_DE_60M,
-        },
     )
