+++ impact/reglementations/tests/conftest.py
@@ -16,7 +16,7 @@ from reglementations.models.csrd import DocumentAnalyseIA
 
 # Empêche tous les tests de faire des appels api
 @pytest.fixture(autouse=True)
-def mock_api(mock_api_recherche_entreprises, mock_api_egapro, mock_api_bges):
+def mock_api(mock_api_infos_entreprise, mock_api_egapro, mock_api_bges):
     pass
